refactor: clean up module scaffolding and remove code duplication
- Delete dead scaffolding: Skeleton/vite.config.js, Skeleton/package.json
- Fix Skeleton/composer.json: remove stale nwidart author/name defaults
- Fix SkeletonDatabaseSeeder: add strict_types, remove commented-out code
- Remove empty register() method from SkeletonServiceProvider
- DRY: extract shared page resolver to lib/resolve-module-page.ts
- DRY: simplify getPermissions() to use collectFromContract()
