+++ Modules/Skeleton/composer.json
@@ -1,18 +1,10 @@
 {
-    "name": "nwidart/skeleton",
-    "description": "",
-    "authors": [
-        {
-            "name": "Nicolas Widart",
-            "email": "n.widart@gmail.com"
-        }
-    ],
+    "name": "modules/skeleton",
+    "description": "Skeleton module for validating the modular architecture infrastructure",
     "extra": {
         "laravel": {
             "providers": [],
-            "aliases": {
-
-            }
+            "aliases": {}
         }
     },
     "autoload": {
