+++ Modules/Skeleton/database/seeders/SkeletonDatabaseSeeder.php
@@ -1,16 +1,12 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Modules\Skeleton\Database\Seeders;
 
 use Illuminate\Database\Seeder;
 
 class SkeletonDatabaseSeeder extends Seeder
 {
-    /**
-     * Run the database seeds.
-     */
-    public function run(): void
-    {
-        // $this->call([]);
-    }
+    public function run(): void {}
 }
