--- Modules/Skeleton/package.json
@@ -1,15 +1,0 @@
-{
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build"
-  },
-  "devDependencies": {
-    "axios": "^1.1.2",
-    "laravel-vite-plugin": "^0.7.5",
-    "sass": "^1.69.5",
-    "postcss": "^8.3.7",
-    "vite": "^4.0.0"
-  }
-}
